% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/followers.R
\name{followers}
\alias{followers}
\title{Followers}
\usage{
followers(token = NULL, username = NULL)
}
\arguments{
\item{token}{Access Token}

\item{username}{Screen Name}
}
\description{
Return ids of followers of an user
}
\examples{
followers(<your access token>, "xyz")

}

