% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{hashTag}
\alias{hashTag}
\title{Hash Tag Twitter.}
\usage{
hashTag(token = NULL, hashtag = NULL, count = 15, lang = "en")
}
\arguments{
\item{token}{Access Token}

\item{hashtag}{Hash Tag}

\item{count}{Maximum number of tweets to return per page,
up to a maximum of 100. Defaults to 15.}

\item{lang}{Restricts tweets to the given language,
given by an ISO 639-1 code}
}
\description{
Search for tweets having particular hashtag.
}
\examples{
hashTag(token, "oldisgold")
hashTag(token, "oldisgold", count = 10)
hashTag(token, "oldisgold", count = 50, lang = "zh")
}

