% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oauth.R
\name{tweetOauth}
\alias{tweetOauth}
\title{Twitter Oauth authentication.}
\usage{
tweetOauth(client_id = NULL, client_secret = NULL)
}
\arguments{
\item{client_id}{API Key}

\item{client_secret}{API Secret}
}
\description{
To get an access token with which further API call is possible.
}

