% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{userInfo}
\alias{userInfo}
\title{User Information.}
\usage{
userInfo(token = NULL, user_name = NULL, user_id = NULL)
}
\arguments{
\item{token}{Access Token}

\item{user_name}{Screen Name}

\item{user_id}{User Id}
}
\description{
General Information of User.
}
\examples{
userInfo(token, "put username here")
}

