% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeline.R
\name{userTweets}
\alias{userTweets}
\title{User Timeline.}
\usage{
userTweets(token = NULL, user_name = NULL, user_id = NULL, count = 15)
}
\arguments{
\item{token}{Access Token}

\item{user_name}{Screen Name}

\item{user_id}{User Id}

\item{count}{Maximum Tweets}
}
\description{
Latest tweets of an user.
}
\examples{
userTweets("your access token", "any valid username")
}

